Move item deletion functions from manage.py to deleting.py to avoid cyclic imports
